refactor(mcp): split SessionService dataset and metrics halves into mixins
Move the dataset-structure/statistics methods to dataset_inspection.py
and the live-metric/leaderboard/curves/compare methods to
metrics_service.py, verbatim; SessionService now inherits both mixins so
every call site and test keeps the same surface.
